蓝盾oauth2鉴权实现 #9110

+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/dao/AuthOauth2CodeDao.kt
@@ -22,7 +22,7 @@ class AuthOauth2CodeDao {
         dslContext: DSLContext,
         code: String,
         clientId: String,
-        expiredTime: Int
+        expiredTime: Long
     ): Int {
         return with(TAuthOauth2Code.T_AUTH_OAUTH2_CODE) {
             dslContext.insertInto(
@@ -33,7 +33,7 @@ class AuthOauth2CodeDao {
             ).values(
                 clientId,
                 code,
-                expiredTime
+                expiredTime.toInt()
             ).execute()
         }
     }
